move: 기존의 skeleton ui 폴더 위치 변경
