+++ src/app/components/SkeletonUi/SkeletonBookDetail.tsx
