+++ src/app/page.tsx
@@ -14,10 +14,10 @@ import Link from "next/link";
 import { useEffect, useState } from "react";
 import ManyTalkRoomBookCard from "./components/Card/MainPageCard/ManyTalkRoomBookCard";
 import TalkRoomCard from "./components/Card/MainPageCard/TalkRoomCard";
-import SkeletonBestSeller from "./components/Card/SkeletonUiCard/SkeletonBestSeller";
-import SkeletonManyTalkRoom from "./components/Card/SkeletonUiCard/SkeletonManyTalkRoom";
-import SkeletonTalkRoomCard from "./components/Card/SkeletonUiCard/SkeletonTalkRoomCard";
 import ResizeImage from "./components/ResizeImage/ResizeImage";
+import SkeletonBestSeller from "./components/SkeletonUi/SkeletonBestSeller";
+import SkeletonManyTalkRoom from "./components/SkeletonUi/SkeletonManyTalkRoom";
+import SkeletonTalkRoomCard from "./components/SkeletonUi/SkeletonTalkRoomCard";
 import RankSwiper from "./components/Swiper/RankSwiper";
 import TalkRoomCardSwiper from "./components/Swiper/TalkRoomCardSwiper";
 import { ThemeMain } from "./components/Theme/Theme";
