+++ src/app/talkroom/related/[isbn]/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import TalkRoomCard from "@/app/components/Card/MainPageCard/TalkRoomCard";
-import SkeletonTalkRoomCard from "@/app/components/Card/SkeletonUiCard/SkeletonTalkRoomCard";
 import Pagination from "@/app/components/Pagination/Pagination";
+import SkeletonTalkRoomCard from "@/app/components/SkeletonUi/SkeletonTalkRoomCard";
 import { ThemeMain } from "@/app/components/Theme/Theme";
 import RecentMakeTalkRoom from "@/assets/img/recent-make-talk-room.svg";
 import { useGetBookRelatedTalkRoom } from "@/hook/reactQuery/book/useGetBookRelatedTalkRoom";
